+++ src/lib/utils/index.ts
@@ -1,3 +1,9 @@
 export * from "./preview-cookies";
 
 export let LocaleConfig: Intl.DateTimeFormatOptions = { dateStyle: "full" };
+
+export function extractYouTubeId(url: string) {
+    const regex = /(?:https?:\/\/)?(?:www\.)?(?:youtube\.com\/(?:[^\/\n\s]+\/\S+\/|(?:v|e(?:mbed)?)\/|\S*?[?&]v=)|youtu\.be\/)([a-zA-Z0-9_-]{11})/;
+    const match = url.match(regex);
+    return match ? match[1] : null;
+}